+++ gradle-plugin/src/main/kotlin/ComposeResourcesPlugin.kt
@@ -1,12 +1,11 @@
 package com.javiersc.compose.resources
 
-import com.android.build.gradle.LibraryExtension
+import com.javiersc.compose.resources.generators.generateComposeResources
+import com.javiersc.compose.resources.utils.androidSourceDirectorySetResources
 import com.javiersc.compose.resources.utils.commonMainKotlin
 import com.javiersc.compose.resources.utils.commonMainResourcesDirs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.findByType
-import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.register
 
 public class ComposeResourcesPlugin : Plugin<Project> {
@@ -17,15 +16,15 @@ public class ComposeResourcesPlugin : Plugin<Project> {
         project.commonMainKotlin?.srcDirs(project.generationDir)
 
         project.commonMainResourcesDirs.forEach { srcDir ->
-            project
-                .extensions
-                .findByType<LibraryExtension>()
-                ?.sourceSets
-                ?.get("main")
-                ?.res
-                ?.srcDir(srcDir)
+            project.androidSourceDirectorySetResources?.srcDir(srcDir)
         }
 
-        project.tasks.register<GenerateComposeResourcesTask>("generateComposeResources")
+        project.afterEvaluate {
+            project.tasks.register<GenerateComposeResourcesTask>(GenerateComposeResourcesTask.name)
+            project.tasks.findByName("preBuild")?.dependsOn(GenerateComposeResourcesTask.name)
+            if (System.getProperty("idea.sync.active")?.toBoolean() == true) {
+                project.generateComposeResources()
+            }
+        }
     }
 }