+++ gradle-plugin/src/main/kotlin/utils/Directories.kt
@@ -1,11 +1,15 @@
 package com.javiersc.compose.resources.utils
 
+import com.android.build.api.dsl.AndroidSourceDirectorySet
+import com.android.build.gradle.LibraryExtension
 import com.javiersc.compose.resources.DefaultSuffix
 import com.javiersc.compose.resources.DrawableDirName
 import com.javiersc.compose.resources.composeResourcesExtension
 import java.io.File
 import org.gradle.api.Project
 import org.gradle.api.file.SourceDirectorySet
+import org.gradle.kotlin.dsl.findByType
+import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.getByType
 import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
@@ -26,6 +30,9 @@ internal val Project.commonMainResources: SourceDirectorySet?
 
 internal val Project.commonMainResourcesDirs: List<File>
     get() = commonMainSourceSet?.resources?.srcDirs?.toList().orEmpty()
+
+internal val Project.androidSourceDirectorySetResources: AndroidSourceDirectorySet?
+    get() = extensions.findByType<LibraryExtension>()?.sourceSets?.findByName("main")?.res
 
 internal val Project.commonMainDrawables: List<File>
     get() =